Struct mtbl::FilesetOptions

pub struct FilesetOptions {
    pub reload_interval_seconds: Option<u32>,
}

Options for opening an MTBL fileset.

Examples

FilesetOptions::new().reload_interval_seconds(10).open_from_path("/tmp/data-fileset")Run

Fields
reload_interval_seconds: Option<u32>

How often, in seconds, to reload the fileset description file to look for new file entries. The mtbl default is 60 seconds.

Methods

impl FilesetOptions

fn new() -> FilesetOptions

Create a FilesetOptions with only defaults.

fn reload_interval_seconds(self: &Self, reload_interval_seconds: u32) -> FilesetOptions

Create a modified FilesetOptions with reload_interval_seconds set.

fn open_from_path<T: AsRef<Path>>(self: &Self, setfile: T, merge_fn: Box<MergeFn>) -> Fileset

Open a Fileset with these options from the specified setfile. Note that you must include a MergeFn to combine colliding entries that have the same key.