Struct mtbl::Merger

pub struct Merger {
    pub merge_fn: Box<Box<MergeFn>>,
    // some fields omitted
}

An MTBL reader that opens and reads from several MTBL files, merging their contents.

An MTBL can have only one value for a key, so when the inputs have a key collision (more than one source contains the same key), it uses a MergeFn to combine them.

Fields
merge_fn: Box<Box<MergeFn>>

The function used to combine values for colliding keys.

Methods

impl Merger

fn merge_choose_last_value(_key: &[u8], _val0: &[u8], val1: &[u8]) -> Vec<u8>

A default MTBL merging function that chooses the last (second) value for the colliding key.

fn merge_choose_first_value(_key: &[u8], val0: &[u8], _val1: &[u8]) -> Vec<u8>

A simple MTBL merging function that chooses the first value for the colliding key.

fn new<F>(sources: Vec<Reader>, merge_fn: F) -> Merger where F: Fn(&[u8], &[u8], &[u8]) -> Vec<u8> + 'static

Create a merger from a collection of other sources. Note that you must provide a merge_fn to combine values for colliding keys.

fn add_source<T: 'static + Read>(self: &mut Self, source: T)

Add an additional source of data to be merged.