Trait mtbl::Read

pub trait Read {
    fn raw_mtbl_source(&self) -> &*const mtbl_source;

    fn get<T>(&self, key: T) -> Option<Vec<u8>> where Self: Sized, T: AsRef<[u8]> { ... }
    fn iter(&self) -> Iter { ... }
    fn get_prefix<T>(&self, prefix: T) -> Iter where Self: Sized, T: AsRef<[u8]> { ... }
    fn get_range<T, U>(&self, key0: T, key1: U) -> Iter where Self: Sized, T: AsRef<[u8]>, U: AsRef<[u8]> { ... }
}

A trait for objects that can read keys from an MTBL file.

A Read is accessed like a sorted map, with each key mapping to one value. Keys and values are both byte sequences, passed around as slices or vectors. Because it's a sorted map, you can access not just via an exact key but also by a key prefix or range.

Required Methods

fn raw_mtbl_source(&self) -> &*const mtbl_source

Get the internal mtbl_source pointer.

Provided Methods

fn get<T>(&self, key: T) -> Option<Vec<u8>> where Self: Sized, T: AsRef<[u8]>

Get the value of a key, if it's present.

fn iter(&self) -> Iter

Get an iterator over all keys and values.

fn get_prefix<T>(&self, prefix: T) -> Iter where Self: Sized, T: AsRef<[u8]>

Get an iterator over all keys and values where the key starts with the given prefix.

fn get_range<T, U>(&self, key0: T, key1: U) -> Iter where Self: Sized, T: AsRef<[u8]>, U: AsRef<[u8]>

Get an iterator over all keys and values, where the keys are between key0 and key1 (inclusive).