Struct mtbl::Fileset

pub struct Fileset {
    pub options: FilesetOptions,
    // some fields omitted
}

An MTBL reader that watches a "setfile" containing a list of MTBL files to read from.

It acts like a Merger that watches that setfile for updates to a list of MTBL files. Note that paths in the setfile are relative paths from the directory of the setfile.

Filesets are not thread-safe because reloading, which happens automatically when reading, is not thread-safe--so only one thread can read from a Fileset.

To create a Fileset with non-default options, see FilesetOptions.

Examples

$ cp my-data.mtbl /tmp/my-data.mtbl
$ echo "my-data.mtbl" >> /tmp/fs.mtbl-fileset
...
let fileset = Fileset::open_from_path("/tmp/fs.mtbl-fileset", my_merge_fn);Run

Fields
options: FilesetOptions

The options used to open this Fileset.

Methods

impl Fileset

fn open_from_path<T: AsRef<Path>>(setfile: T, merge_fn: Box<MergeFn>) -> Fileset

Open a Fileset from a path. Note that you must include a MergeFn to combine colliding entries (entries that have the same key).

fn reload(&mut self)

Reload the list of MTBL files (ignored if less than the configured reload interval has passed).