Struct mtbl::Reader

pub struct Reader {
    pub options: ReaderOptions,
    // some fields omitted
}

A reader for a single MTBL file.

Reader uses a memory-mapped file and is immutable and entirely thread-safe.

To create a Reader with options other than the default, use ReaderOptions.

Fields

options

The options used to open this MTBL file.

Methods

impl Reader

fn open_from_path<T: AsRef<Path>>(path: T) -> IOResult<Reader>

Open an MTBL reader from a file described by the given path.

fn open_from_file<T: 'static + AsRawFd>(file: &T) -> IOResult<Reader>

Open an MTBL reader from a file object.